+++ src/main/java/racingcar/controller/WebController.java
@@ -2,7 +2,6 @@ package racingcar.controller;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -10,10 +9,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import racingcar.dto.CarDTO;
 import racingcar.dto.ResultDTO;
-import racingcar.dto.RacingGameRequestDTO;
-import racingcar.dto.RacingGameResponseDTO;
+import racingcar.request.RacingGameRequest;
+import racingcar.response.RacingGameResponse;
 import racingcar.exception.CarException;
 import racingcar.exception.CarsException;
 import racingcar.service.RacingCarService;
@@ -30,26 +28,26 @@ public class WebController {
 
     @PostMapping("/plays")
     @ExceptionHandler(value = {CarException.class, CarsException.class})
-    public RacingGameResponseDTO createGame(@RequestBody RacingGameRequestDTO racingGameRequestDTO) {
-        final List<String> carNames = Arrays.stream(racingGameRequestDTO.getNames().split(DELIMITER))
+    public RacingGameResponse createGame(@RequestBody RacingGameRequest racingGameRequest) {
+        final List<String> carNames = Arrays.stream(racingGameRequest.getNames().split(DELIMITER))
                 .collect(Collectors.toList());
-        final ResultDTO resultDTO = racingCarService.play(carNames, racingGameRequestDTO.getCount());
+        final ResultDTO resultDTO = racingCarService.play(carNames, racingGameRequest.getCount());
         final String winnerNames = resultDTO.getWinners().stream()
                 .collect(Collectors.joining(DELIMITER));
-        return new RacingGameResponseDTO(winnerNames, resultDTO.getCarDTOs());
+        return new RacingGameResponse(winnerNames, resultDTO.getCarDTOs());
     }
 
     @GetMapping("/plays")
-    public List<RacingGameResponseDTO> getGameResults() {
-        final List<RacingGameResponseDTO> racingGameResponseDTOs = new ArrayList<>();
+    public List<RacingGameResponse> getGameResults() {
+        final List<RacingGameResponse> racingGameResponses = new ArrayList<>();
         final List<ResultDTO> resultDTOs = racingCarService.getSavedGames();
 
         for (ResultDTO resultDTO : resultDTOs) {
             String winnerNames = resultDTO.getWinners()
                     .stream()
                     .collect(Collectors.joining(DELIMITER));
-            racingGameResponseDTOs.add(new RacingGameResponseDTO(winnerNames, resultDTO.getCarDTOs()));
+            racingGameResponses.add(new RacingGameResponse(winnerNames, resultDTO.getCarDTOs()));
         }
-        return new ArrayList<>(racingGameResponseDTOs);
+        return new ArrayList<>(racingGameResponses);
     }
 }
