+++ src/main/java/racingcar/response/RacingGameResponse.java
@@ -1,16 +1,17 @@
-package racingcar.dto;
+package racingcar.response;
 
 import java.util.List;
+import racingcar.dto.CarDTO;
 
-public class RacingGameResponseDTO {
+public class RacingGameResponse {
 
     private String winners;
     private List<CarDTO> racingCars;
 
-    public RacingGameResponseDTO() {
+    public RacingGameResponse() {
     }
 
-    public RacingGameResponseDTO(final String winners, final List<CarDTO> racingCars) {
+    public RacingGameResponse(final String winners, final List<CarDTO> racingCars) {
         this.winners = winners;
         this.racingCars = racingCars;
     }
